[APP-2865] Move ApkfyViewModel to feature-apkfy module
