+++ feature-apkfy/build.gradle.kts
@@ -11,5 +11,7 @@ android {
 
 dependencies {
   implementation(project(ModuleDependency.APTOIDE_NETWORK))
+  implementation(project(ModuleDependency.EXTENSIONS))
+  implementation(project(ModuleDependency.FEATURE_APPS))
   implementation(LibraryDependency.DATASTORE)
 }
